+++ backend/Cassino.Application/DependencyInjection.cs
@@ -35,6 +35,7 @@ public static class DependencyInjection
             .AddScoped<IAuthService, AuthService>()
             .AddScoped<INotificator, Notificator>()
             .AddScoped<IUsuarioService, UsuarioService>()
+            .AddScoped<ISaldoService, SaldoService>()
             .AddScoped<IUsuarioAuthService, UsuarioAuthService>()
             .AddScoped<IPasswordHasher<Administrador>, Argon2PasswordHasher<Administrador>>()
             .AddScoped<IPasswordHasher<Usuario>, Argon2PasswordHasher<Usuario>>();
